Swap self for &self
