+++ src/vm/runners/builtin_runner/bitwise.rs
@@ -15,8 +15,6 @@ use crate::vm::vm_core::VirtualMachine;
 use crate::vm::vm_memory::memory::Memory;
 use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
 
-use super::BuiltinRunner;
-
 pub struct BitwiseBuiltinRunner {
     _ratio: u32,
     pub base: isize,
@@ -119,8 +117,18 @@ impl BitwiseBuiltinRunner {
         ("bitwise", (self.base, self.stop_ptr))
     }
 
+    pub fn get_used_cells(&self, vm: &VirtualMachine) -> Result<usize, MemoryError> {
+        let base = self.base();
+        vm.segments
+            .get_segment_used_size(
+                base.try_into()
+                    .map_err(|_| MemoryError::AddressInTemporarySegment(base))?,
+            )
+            .ok_or(MemoryError::MissingSegmentUsedSizes)
+    }
+
     pub fn get_used_cells_and_allocated_size(
-        self,
+        &self,
         vm: &VirtualMachine,
     ) -> Result<(usize, usize), MemoryError> {
         let ratio = self._ratio as usize;
@@ -129,8 +137,7 @@ impl BitwiseBuiltinRunner {
         if vm.current_step < min_step {
             Err(MemoryError::InsufficientAllocatedCells)
         } else {
-            let builtin = BuiltinRunner::Bitwise(self);
-            let used = builtin.get_used_cells(vm)?;
+            let used = self.get_used_cells(vm)?;
             let size = (cells_per_instance
                 * safe_div(&bigint!(vm.current_step), &bigint!(ratio))
                     .map_err(|_| MemoryError::InsufficientAllocatedCells)?)
@@ -368,8 +375,7 @@ mod tests {
 
     #[test]
     fn get_used_cells_missing_segment_used_sizes() {
-        let builtin =
-            BuiltinRunner::Bitwise(BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default()));
+        let builtin = BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default());
         let vm = vm!();
 
         assert_eq!(
@@ -380,8 +386,7 @@ mod tests {
 
     #[test]
     fn get_used_cells_empty() {
-        let builtin =
-            BuiltinRunner::Bitwise(BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default()));
+        let builtin = BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default());
         let mut vm = vm!();
 
         vm.segments.segment_used_sizes = Some(vec![0]);
@@ -390,8 +395,7 @@ mod tests {
 
     #[test]
     fn get_used_cells() {
-        let builtin =
-            BuiltinRunner::Bitwise(BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default()));
+        let builtin = BitwiseBuiltinRunner::new(&BitwiseInstanceDef::default());
         let mut vm = vm!();
 
         vm.segments.segment_used_sizes = Some(vec![4]);
