+++ src/vm/runners/builtin_runner/ec_op.rs
@@ -15,8 +15,6 @@ use crate::vm::vm_memory::memory::Memory;
 use crate::vm::vm_memory::memory_segments::MemorySegmentManager;
 use crate::{bigint, bigint_str};
 
-use super::BuiltinRunner;
-
 pub struct EcOpBuiltinRunner {
     _ratio: u32,
     pub base: isize,
@@ -203,8 +201,18 @@ impl EcOpBuiltinRunner {
         ("ec_op", (self.base, self.stop_ptr))
     }
 
+    pub fn get_used_cells(&self, vm: &VirtualMachine) -> Result<usize, MemoryError> {
+        let base = self.base();
+        vm.segments
+            .get_segment_used_size(
+                base.try_into()
+                    .map_err(|_| MemoryError::AddressInTemporarySegment(base))?,
+            )
+            .ok_or(MemoryError::MissingSegmentUsedSizes)
+    }
+
     pub fn get_used_cells_and_allocated_size(
-        self,
+        &self,
         vm: &VirtualMachine,
     ) -> Result<(usize, usize), MemoryError> {
         let ratio = self._ratio as usize;
@@ -213,8 +221,7 @@ impl EcOpBuiltinRunner {
         if vm.current_step < min_step {
             Err(MemoryError::InsufficientAllocatedCells)
         } else {
-            let builtin = BuiltinRunner::EcOp(self);
-            let used = builtin.get_used_cells(vm)?;
+            let used = self.get_used_cells(vm)?;
             let size = (cells_per_instance
                 * safe_div(&bigint!(vm.current_step), &bigint!(ratio))
                     .map_err(|_| MemoryError::InsufficientAllocatedCells)?)
@@ -802,7 +809,7 @@ mod tests {
 
     #[test]
     fn get_used_cells_missing_segment_used_sizes() {
-        let builtin = BuiltinRunner::EcOp(EcOpBuiltinRunner::new(&EcOpInstanceDef::default()));
+        let builtin = EcOpBuiltinRunner::new(&EcOpInstanceDef::default());
         let vm = vm!();
 
         assert_eq!(
@@ -813,7 +820,7 @@ mod tests {
 
     #[test]
     fn get_used_cells_empty() {
-        let builtin = BuiltinRunner::EcOp(EcOpBuiltinRunner::new(&EcOpInstanceDef::default()));
+        let builtin = EcOpBuiltinRunner::new(&EcOpInstanceDef::default());
         let mut vm = vm!();
 
         vm.segments.segment_used_sizes = Some(vec![0]);
@@ -822,7 +829,7 @@ mod tests {
 
     #[test]
     fn get_used_cells() {
-        let builtin = BuiltinRunner::EcOp(EcOpBuiltinRunner::new(&EcOpInstanceDef::default()));
+        let builtin = EcOpBuiltinRunner::new(&EcOpInstanceDef::default());
         let mut vm = vm!();
 
         vm.segments.segment_used_sizes = Some(vec![4]);
